simplify for nais admin usecase
since this is a tool mainly for our own use, we can simplify the interface a bit and be more opinionated
another point is that we should all have the same context names, so that assumptions made in e.g. Makefiles
for local development is the same for all of us.